Separate the Navigator component
